[zeroD] Move newWall to ConnectorFactory
fix doxygen for flowdeviceGroup
